+++ src/c_cmds.c
@@ -1503,15 +1503,17 @@ static void C_UnbindDuplicates(const int keep, const controltype_t type, const i
             {
                 if (actions[i].keyboard1 && controls[control].value == *(int *)actions[i].keyboard1)
                 {
-                    C_Warning(1, "The " BOLD("%s") " control may only be bound to one action. The " BOLD("%s")
-                        " action has been unbound.", controls[control].control, actions[i].action);
+                    C_Warning(1, "Controls may only be bound to one action. The duplicate " BOLD("%s")
+                        " action has been unbound from the " BOLD("%s")" control.",
+                        actions[i].action, controls[control].control);
                     *(int *)actions[i].keyboard1 = 0;
                 }
 
                 if (actions[i].keyboard2 && controls[control].value == *(int *)actions[i].keyboard2)
                 {
-                    C_Warning(1, "The " BOLD("%s") " control may only be bound to one action. The " BOLD("%s")
-                        " action has been unbound.", controls[control].control, actions[i].action);
+                    C_Warning(1, "Controls may only be bound to one action. The duplicate " BOLD("%s")
+                        " action has been unbound from the " BOLD("%s")" control.",
+                        actions[i].action, controls[control].control);
                     *(int *)actions[i].keyboard2 = 0;
                 }
             }
@@ -1519,8 +1521,9 @@ static void C_UnbindDuplicates(const int keep, const controltype_t type, const i
             {
                 if (actions[i].mouse1 && controls[control].value == *(int *)actions[i].mouse1)
                 {
-                    C_Warning(1, "The " BOLD("%s") " control may only be bound to one action. The " BOLD("%s")
-                        " action has been unbound.", controls[control].control, actions[i].action);
+                    C_Warning(1, "Controls may only be bound to one action. The duplicate " BOLD("%s")
+                        " action has been unbound from the " BOLD("%s")" control.",
+                        actions[i].action, controls[control].control);
                     *(int *)actions[i].mouse1 = -1;
                 }
             }
@@ -1528,15 +1531,17 @@ static void C_UnbindDuplicates(const int keep, const controltype_t type, const i
             {
                 if (actions[i].gamecontroller1 && controls[control].value == *(int *)actions[i].gamecontroller1)
                 {
-                    C_Warning(1, "The " BOLD("%s") " control may only be bound to one action. The " BOLD("%s")
-                        " action has been unbound.", controls[control].control, actions[i].action);
+                    C_Warning(1, "Controls may only be bound to one action. The duplicate " BOLD("%s")
+                        " action has been unbound from the " BOLD("%s")" control.",
+                        actions[i].action, controls[control].control);
                     *(int *)actions[i].gamecontroller1 = 0;
                 }
 
                 if (actions[i].gamecontroller2 && controls[control].value == *(int *)actions[i].gamecontroller2)
                 {
-                    C_Warning(1, "The " BOLD("%s") " control may only be bound to one action. The " BOLD("%s")
-                        " action has been unbound.", controls[control].control, actions[i].action);
+                    C_Warning(1, "Controls may only be bound to one action. The duplicate " BOLD("%s")
+                        " action has been unbound from the " BOLD("%s")" control.",
+                        actions[i].action, controls[control].control);
                     *(int *)actions[i].gamecontroller2 = 0;
                 }
             }